requirements.txt
absl-py
easydict
cython
numpy==1.16.4
tensorflow-gpu==1.14.0
tqdm
